requirements.txt
Flask
Pillow
numpy
scikit-learn
joblib
matplotlib
